Update_config.yml
Let's change the _config.yml so it's a perfect fit for your new blog. First, we need to use a blog-ready theme. For this activity, we will use a theme named minima.

Navigate to the Code tab of this repository, and browse to the _config.yml file, or click this link here.
In the upper right corner, click ✏️ to open the file editor.
Add a theme: set to minima so it shows in the _config.yml file as below:
theme: minima
Modify the other configuration variables such as title:, author:, and description: to customize your site.
Click Create a new branch for this commit and start a pull request.
Open a pull request.